==> R/00_doit.R <==
# Run all scripts ---------------------------------------------------------
source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_analysis_i.R")

source(file = "R/01_load_FCS.R")
source(file = "R/02_clean_FCS.R")
source(file = "R/03_augment_FCS.R")
source(file = "R/04_analysis_FCS_i.R")
